requirements.txt
langchain
chromadb
sentence-transformers
pypdf
streamlit
python-dotenv
huggingface_hub
